WIP trying to use distribution file as generic rosdep source.
I spent about 90 minutes trying to set this up and I have learned a few
things.

Distribution files are automatically pulled in, converted, and pickled
in the cache so the code in this commit is actually somewhat
redundant... at least it seems so.
If nothing else, the cache filenames collide so unless you munge the
filename, which is coming in the next commit, you don't actually add any
new data to the rosdep cache.

Similarly when trying to resolve a key you would expect from a
distribution file, it does not work. I've been digging at this and all I
can tell so far is that it has something to do with rosdep limiting
input based on what files it thinks are appropriate for the current
environment.

I think the next place to look is in how the default dependency lists
are assembled and we need to look for why my kinetic-generated source
isn't on that list.