Prepend rosdistro+ to the source url for rosdistro sources.
rosdistro files are pulled on each rosdep update and then only included
selectively (which means we might be able to change the behavior here
from re-creating that functionality to enabling keys from a parent
distro in an addon distro) but because of this the cached rosdep files
for the regular distribution.yaml and our added one collide in the
cache. So add something foolish to the hashed string to distinguish ours
from the mainline one.